Componentize some dashboard
